+++ src/main/kotlin/com/barostartbe/domain/example/repository/ExampleRepository.kt
@@ -1,6 +1,6 @@
 package com.barostartbe.domain.example.repository
 
-import com.barostartbe.domain.example.domain.Example
+import com.barostartbe.domain.example.entity.Example
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface ExampleRepository : JpaRepository<Example, Long> {
